Implemented drone enemy
I implemented a drone enemy using the sprite I created earlier. The drone follows a path then when a player gets close it aborts it and chases the player. Code is similar to that of the patrolling object in the workshop we had. I added variable definitions for many of the attributes for easy editing. I also increased the jump height of the player. Damage still needs to be implemented.